feat!: Comment out all SCSS and clean HTML to start from a clean slate
Index.html is the new fundament to restart out efforts to build a rich and efficient website.
This version needs to be controlled, checked and worked on before it is merged with main.

BREAKING CHANGE!: This will render almost all previous tickets and issues to zero as we are starting "fresh".